Extend secrets to be 48 characters long
Without a length annotation, I got this fatal error
from the proxy pod (as the other pods CrashLooped, too):

    error creating proxy service: proxy: invalid 'SHARED_SECRET': cryptutil: got 30 bytes but want 32

Considering that the default generated secret length is 40, I guess this
means the sharedSecret is processed as Base64 recursively?
Anyway, this ought to give it the data size it's looking for.