Remove hats subsystem
On further inspection, I don't think this ever worked (vcgencmd not working in container), and even if it did, all it would do is slow the fan down, which isn't something I need or even really want.

The node this hat's on has been having reliability problems, and dropping this gives us one less possible point of failure, and one less potential vulnerability in a pod with privileged context.